requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
PyPDF2==3.0.1
pypdf==4.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2
openai>=1.10.0,<2.0.0
